refactor: RuleNotifyButton 버튼 분리 (#13)
